+++ changelog.md
@@ -2,6 +2,7 @@ GDB+ 7.1.1
 * Check libc.address and elf.address correctly aligned.
 * allow Debugger(silent=True) to hide pwntools default logs
 * Execute from_entry only if inside the loader
+* Allow user to access all libraries as they would with the libc
 
 GDB+ 7.1.0 QEMU support
 * Find libc used by process under QEMU